2.5更新
基本完善通过列表选择城市功能
修正bug：
1. 天气json数据类型错误
2. 列表显示逻辑

+++ app/src/main/java/com/example/nbdv/weatherdemo/Weather.java
@@ -108,7 +108,7 @@ class Now
     public  Cond cond;
     public int fl;
     public int hum;
-    public int pcpn;
+    public float pcpn;
     public int pres;
     public int tmp;
     public  int vis;
